pyproject.toml
[tool.poetry]
name = "filesystemui"
version = "0.1.0"
description = ""
authors = ["Daniel"]

[tool.poetry.dependencies]
python = "^3.10"
mypy = "^1.4.1"
Pillow = "^10.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.mypy]
strict = true